Move MetricsSource to a distinct package to reduce circular dependencies
