+++ ts/libs/core/src/index.ts
@@ -9,6 +9,7 @@ export * from './lib/model';
 export * from './lib/orchestrator/public';
 export * from './lib/raw-metrics-query/public';
 export * from './lib/composed-metrics/public';
+export * from './lib/metrics/public';
 export * from './lib/elasticity/public';
 export * from './lib/slo/public';
 export * from './lib/transformation/public';
@@ -20,6 +21,7 @@ export * from './lib/orchestrator/public/impl';
 export * from './lib/raw-metrics-query/public/time-series/execution/impl';
 export * from './lib/composed-metrics/public/common/impl';
 export * from './lib/composed-metrics/public/control/impl';
+export * from './lib/metrics/public/impl';
 export * from './lib/elasticity/public/control/impl';
 export * from './lib/elasticity/public/service/impl';
 export * from './lib/slo/public/control/impl';
