+++ ts/libs/core/src/lib/metrics/public/impl/index.ts
@@ -1,0 +1,1 @@
+export * from './default-metrics-sources-manager';
