+++ ts/libs/core/src/lib/runtime/public/impl/index.ts
@@ -1,3 +1,2 @@
-export * from './default-metrics-sources-manager';
 export * from './default-microcontroller-factory';
 export * from './polaris-runtime.base';
