+++ ts/libs/core/src/lib/runtime/public/impl/polaris-runtime.base.ts
@@ -3,6 +3,8 @@ import { DefaultComposedMetricsManager } from '../../../composed-metrics/public/
 import { ElasticityStrategyManager, ElasticityStrategyService } from '../../../elasticity';
 import { DefaultElasticityStrategyManager } from '../../../elasticity/public/control/impl/default-elasticity-strategy-manager';
 import { DefaultElasticityStrategyService } from '../../../elasticity/public/service/impl/default-elasticity-strategy.service';
+import { MetricsSourcesManager } from '../../../metrics';
+import { DefaultMetricsSourcesManager } from '../../../metrics/public/impl';
 import { ObjectKindWatcher, OrchestratorClient, WatchManager } from '../../../orchestrator';
 import { DefaultWatchManager } from '../../../orchestrator/public/impl';
 import { SloControlLoop, SloEvaluator } from '../../../slo';
@@ -10,9 +12,7 @@ import { DefaultSloControlLoop } from '../../../slo/public/control/impl';
 import { PropertyTransformer } from '../../../transformation/internal/property-transformer';
 import { PolarisTransformationServiceManager } from '../../../transformation/public/common';
 import { DefaultPolarisTransformationService } from '../../../transformation/public/impl';
-import { MetricsSourcesManager } from '../metrics-source';
 import { PolarisRuntime } from '../polaris-runtime';
-import { DefaultMetricsSourcesManager } from './default-metrics-sources-manager';
 import { DefaultMicrocontrollerFactory } from './default-microcontroller-factory';
 
 /**
