+++ ts/libs/core/src/lib/runtime/public/index.ts
@@ -1,3 +1,2 @@
-export * from './metrics-source';
 export * from './microcontroller-factory';
 export * from './polaris-runtime';
